bootcamp/week-11-testing_seguridad_docker/3-proyecto/starter/pyproject.toml
[project]
name = "week-11-library-server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "mcp==1.9.0",
    "aiosqlite==0.20.0",
    "httpx==0.27.2",
    "pydantic==2.11.4",
    "python-dotenv==1.0.1",
]

[dependency-groups]
test = [
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "pytest-cov==6.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
